Publish to PyPI with trusted publisher
This commit will replace the `hatch publish` command from the template
repo with a PyPI trusted publisher.

The advantage to OIDC is that authentication is performed with temporary
API tokens (only valid for 15 minutes) instead of persistent tokens that
must be manually generated on PyPI and pasted into GitHub Secrets. The
disadvantage is that authentication is more complicated.

To use PyPI OIDC, a
[trusted publisher](https://docs.pypi.org/trusted-publishers/) is set up
for the PyPI project. Next, a dedicated GitHub Actions deployment
environment is created for PyPI with protection rules that only allow
use of the environment with Git tags. The environment protection rules
can be combined with tag protection rules in GitHub rulesets to ensure
PyPI packages can only be published if a maintainer triggers a workflow
run with a Git tag ref.

The GitHub Actions workflow will be updated to use the deployment
environment. Deployment environments must be selected at the job level
before the job begins, so a setup job will be added that selects the
appropriate deployment environment and passes it to the PyPI job. Each
use of a deployment environment creates a deployment that can be either
active or inactive. GitHub Actions auto-inactivates deployments, and
although this behavior is not configurable or documented, one workaround
is to provide each deployment with its own unique URL.

To publish the Python package to PyPI, the build job will output package
build files to `dist/` and upload them with actions/upload-artifact. The
PyPI job will then download them with actions/download-artifact and use
pypa/gh-action-pypi-publish to upload to PyPI.