Implement `suggest` subcommand
This commit will provide a minimal working implementation of a `suggest`
subcommand. The purpose of the `suggest` subcommand is to suggest
dimensions for cropping images of the given image type.

Images often need to be cropped to specific aspect ratios and dimensions
for upload to sites like TheMovieDB
(https://www.themoviedb.org/bible/image).

- Backdrops: 16:9 (minimum 1280x720 pixels, maximum 3840x2160 pixels)
- Posters: 2:3 (minimum 500x750 pixels, maximum 2000x3000 pixels)

This command suggests dimensions to use for cropping. Examples:

```sh
crp suggest --width=3940 --height 2160 backdrop -> Crop to 3840x2160
crp suggest --width 1652 --height 2214 poster -> Crop to 1476x2214
```

`-h` is not used as a short option for `--height` because it would
conflict with the `-h` used for help.

This command does not currently enforce minimum and maximum dimensions.

`enum.StrEnum` was introduced in Python 3.11, so this commit will set
the minimum Python version to 3.11.
https://docs.python.org/3/library/enum.html