Enforce minimum and maximum image dimensions
This commit will add enforcement of minimum and maximum image dimensions
to the `suggest` subcommand.

- Backdrops (16:9): minimum 1280x720 pixels, maximum 3840x2160 pixels
- Posters (2:3): minimum 500x750 pixels, maximum 2000x3000 pixels

This will be accomplished with a dataclass that checks image dimensions
as part of its `__init__` method. Tests will be added to verify that the
error messages and exit codes are propagated back to the user.