Fix ref in GitHub Actions changelog job
Commit 755610f removed references to the unnecessary `develop` branch.
One of the references that was removed was the `ref: develop` input to
`actions/checkout`. This should have been changed to `ref: main`, but it
was removed entirely. As a result, the `peter-evans/create-pull-request`
step in the changelog job raised an error:

> When the repository is checked out on a commit instead of a branch,
> the 'base' input must be supplied.

Rather than supply the `base` input, this commit will set `ref: main` on
`actions/checkout`.