Add `workflow_dispatch` to changelog job
This commit will add a `workflow_dispatch` manual trigger to the GitHub
Actions job that generates the changelog. This can be helpful if a job
failed when it was triggered by a tag. A new job can be triggered with
`workflow_dispatch` after making subsequent commits.

The changelog job depends on the PyPI job, but the PyPI job will only
run on Git tags. GitHub Actions doesn't really allow "conditional needs"
(using `if:` conditionals with `needs:` so that `needs:` can be skipped)
but it is possible to control the job with deployment environments. This
commit will remove `pypi` from `needs:` but will add a dependency on a
deployment environment. Administrators can use `workflow_dispatch` with
the appropriate deployment environment to trigger the changelog job.