Drop support for Python 3.11
This commit will drop support for Python 3.11 and set the minimum
required version to Python 3.12. This is needed due to the use of the
`@typing.override` decorator that was introduced in Python 3.12. It is
also possible to install the `typing_extensions` package as a runtime
dependency if support for older versions of Python is needed. This
project is new and does not need to support older versions of Python, so
it is simpler to just drop support for Python 3.11.

Users who are invoking version 0.0.2 of the package will see an error:
`ModuleNotFoundError: No module named 'typing_extensions'`. The
`typing_extensions` package is not in the standard library, so it must
be specified in the `pyproject.toml` requirements file and installed.

It's unclear why this error was not seen during development or in any of
the GitHub Actions jobs (even the 3.11 job). It's also confusing because
the typing docs say, "There is no runtime checking of this property."
That statement almost makes it sound like the `@typing.override`
decorator behaves like a type annotation, but that is not the case. For
example, this code:

```py
import dataclasses

if TYPE_CHECKING:
    import sys

    if sys.version_info < (3, 12):
        from typing_extensions import override
    else:
        from typing import override

@dataclasses.dataclass
class AspectRatio:
    width: int
    height: int

    @override
    def __str__(self) -> str:
        return f"{self.width}:{self.height}"
```

Will raise `NameError: name 'override' is not defined` at runtime. So
although it is not _checked_ at runtime, it is _required_ at runtime.

https://docs.python.org/3/library/typing.html#typing.override