Don't use same deployment environment name twice
GitHub Deployment Environments strike again. As mentioned in 5012782
when configuring the PyPI trusted publisher workflow, each use of a
deployment environment creates a deployment that can be either active or
inactive. If two deployments have the same URL, GitHub Actions will
consider the first one (the earlier one) to be active, will set the
second one (the later one) to inactive, and will show the deployment
environment on the homepage as inactive. Although this behavior is not
configurable or documented, one workaround is to provide each deployment
with its own unique URL.

When the `workflow_dispatch` trigger was added to the changelog job in
e8d926e, the job was updated to require a deployment environment as yet
another workaround for how GitHub Actions lacks "conditional needs."
This meant that now potentially two jobs were using the same deployment
URL (the PyPI job and the changelog job), so deployments are now showing
up as inactive.

This commit will hard-code the changelog job to using the Production
environment. Deploying to Production with the PyPI URL will not clash
with the URLs that Vercel uses and should hopefully avoid inactivating
deployments.

https://docs.github.com/en/actions/how-tos/deploy/configure-and-manage-deployments/manage-environments