Switch type checking from mypy to BasedPyright
This commit will switch Python type checking from mypy to
[BasedPyright](https://docs.basedpyright.com/latest/). See the
[docs](https://docs.basedpyright.com/latest/usage/mypy-comparison/) for
a comparison with mypy.

In addition to superior type checking results, another motivation for
this change is the language server provided by the BasedPyright
VSCode/VSCodium extension.
[Jedi](https://github.com/pappasam/jedi-language-server) is no longer
working well for
[Python editing](https://code.visualstudio.com/docs/python/editing)
features like autocompletion and go to definition, and the Microsoft
Pylance/Pyright extension is only available on the proprietary Microsoft
extension registry. BasedPyright is another option also available on
[Open VSX](https://open-vsx.org/extension/detachhead/basedpyright) for
users of [VSCodium](https://code.visualstudio.com/docs/python/editing).

Note that the VSCode `settings.json` for the user should include
`"python.languageServer": "None"` so that BasedPyright can act as the
language server. `"None"` is currently the default for this setting, but
it could change in the future, so it is prudent to explicitly set it to
`"None"`.