Move contributing symlink to `.github`
`CONTRIBUTING.md` is just a symlink. This commit will move the symlink
to `.github` to make the repo root a little cleaner. GitHub will still
pick it up and display a tab above the README. Note that GitHub does not
do the same thing with `CHANGELOG.md`, so the `CHANGELOG.md` symlink
should stay at the repo root for visibility.