Revert "Move contributing symlink to `.github`"
This reverts commit 258b924047544d1b6da951eed154a49200075a6e.

Weirdly, GitHub does not handle the `CONTRIBUTING.md` symlink the same
way whether it is at the repo root or in `.github`. At the repo root,
clicking the "Contributing" tab above the README shows the rendered
Markdown. However, if the `CONTRIBUTING.md` symlink is in `.github`,
GitHub just displays the symlink destination.