Handle missing dimensions in `suggest` subcommand
The Click docs on arguments explain:

> It is possible to make an argument required by setting
> `required=True`. It is not recommended since we think command line
> tools should gracefully degrade into becoming no ops. We think this
> because command line tools are often invoked with wildcard inputs and
> they should not error out if the wildcard is empty.

https://click.palletsprojects.com/en/stable/arguments/

In the spirit of keeping arguments optional, the `width` and `height`
arguments to the `crp suggest` subcommand are both optional. However,
with the code as it is now, both of these arguments should be required.
Without either width or height, the command will raise the exception
`'<' not supported between instances of 'NoneType' and 'int'` which
originates from `if width < minimum_width or height < minimum_height` in
`crp.types.Image.__init__` (Click sets the missing value to `None`).

Missing dimensions will now be handled in the following manner:

- If neither width nor height are supplied, raise an exception.
- If only width or only height is supplied, set the missing dimension to
  a default value appropriate for the aspect ratio of the image type.